--- src/components/utils/Poolable.js
@@ -1,14 +1,0 @@
-Crafty.c("Poolable", {
-  init() {
-    this.attr({ hasPool: true });
-  },
-
-  setupPool(pool) {
-    this.pool = pool;
-    return this;
-  },
-
-  recycle() {
-    this.pool.recycle(this);
-  }
-});
