+++ src/components/utils/index.js
@@ -1,4 +1,3 @@
 import "./flipable";
 import "./hideable";
 import "./Scalable";
-import "./WaterSplashes";
